feat: sidebar redesign, element persistence, viewer routing prep, re-enable APIs
Sidebar:
- Split ProjectWorkspaceView into 16 sub-components under components/workspace/components/
- ELEMENTS panel always visible; localStorage persistence for created elements (ws-elements-<id>)
- DRAWINGS as collapsible bottom drawer that slides up over elements (height animation)
- AssignItemsModal shows real created elements instead of mock data
- BBSEntryModal: max-w-3xl, both buttons h-10
- CreateNewElementModal passes element data back via onCreate callback

Auth and APIs re-enabled:
- AuthGuard: BYPASS_AUTH set to false
- useGetProjectsQuery restored in solo and enterprise project pages
- useGetProjectByIdQuery restored in project layout, WorkspaceLayout, WorkspaceView
- useGetClientsQuery restored in StepProjectDetails
- useCreateProjectMutation restored in ManualSetupShell

Types:
- Added CreatedElement to types.ts
- VIEWER_MAP groundwork for multi-format viewer routing (IFC, DXF, FBX, OBJ, etc.)
